chore: refactored folder structure, renaming Modules to features in resources/js for inertia views
