+++ resources/js/Features/Properties/BulkActions/BulkActionSelector.tsx
