+++ resources/js/Features/Properties/BulkActions/BulkDeleteProperties.tsx
