+++ resources/js/Features/Properties/DeleteProperty.tsx
