+++ resources/js/Features/Properties/PropertyImageGallery.tsx
