+++ resources/js/Features/Properties/PropertyStats.tsx
