+++ resources/js/Features/Properties/PropertyView/LegalFinancialInfo.tsx
