+++ resources/js/Features/Properties/PropertyView/PropertyDetails.tsx
