+++ resources/js/Features/Properties/PropertyView/PropertyMedia.tsx
