+++ resources/js/Features/Properties/PropertyView/PropertySpecifications.tsx
