+++ resources/js/Features/Properties/PropertyView/PropertyView.tsx
