+++ resources/js/Features/Properties/PropertyView/QuickFacts.tsx
