+++ resources/js/Features/Properties/PropertyView/index.ts
