+++ resources/js/Features/Properties/SaveProperty/AssetsTab.tsx
