+++ resources/js/Pages/Properties/Create.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import { router, useForm } from "@inertiajs/react";
 import { Typography, message } from "antd";
 import { Property } from "@/Types";
-import PropertyForm from "@/Modules/Properties/SaveProperty/PropertyForm";
+import PropertyForm from "@/Features/Properties/SaveProperty/PropertyForm";
 
 // Import the new PropertyForm component
